code cleanup: removing unnecessary casts and unnecessary semicolons (trying out static code analysis tools by IntelliJ IDEA)
git-svn-id: file:///srv/local/data/cleanRepo/trunk@24656 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/main/java/org/matsim/utils/objectattributes/attributeconverters/DoubleConverter.java
@@ -32,6 +32,6 @@ public class DoubleConverter implements AttributeConverter<Double> {
 
 	@Override
 	public String convertToString(Object o) {
-		return ((Double) o).toString();
+		return o.toString();
 	}
 }
